requirements.txt
django
torch
pytesseract
pillow
transformers
opencv-python
python-decouple
